+++ lib/active_record_encryption/type.rb
@@ -5,6 +5,7 @@ module ActiveRecordEncryption
     using(ActiveRecordEncryption::SerializerWithCast)
 
     delegate :type, :cast, to: :subtype
+    delegate :user_input_in_time_zone, to: :subtype # for ActiveRecord::AttributeMethods::TimeZoneConversion::TimeZoneConverter
 
     def initialize(subtype: ActiveRecord::Type.default_value, **options)
       # Lookup encryptor from options[:encryption]
@@ -25,6 +26,11 @@ module ActiveRecordEncryption
     def serialize(value)
       serialized = subtype.serialize(value)
       binary.serialize(encryptor.encrypt(serialized)) unless serialized.nil?
+    end
+
+    def changed_in_place?(raw_old_value, value)
+      old_value = deserialize(raw_old_value)
+      @subtype.changed_in_place?(old_value, value)
     end
 
     private